[internal_temperature] Move code into platform specific files
