+++ esphome/components/internal_temperature/internal_temperature.h
@@ -1,14 +1,16 @@
 #pragma once
 
-#include "esphome/core/component.h"
 #include "esphome/components/sensor/sensor.h"
+#include "esphome/core/component.h"
 
 namespace esphome {
 namespace internal_temperature {
 
 class InternalTemperatureSensor : public sensor::Sensor, public PollingComponent {
  public:
+#if defined(USE_ESP32) || (defined(USE_ZEPHYR) && defined(USE_NRF52))
   void setup() override;
+#endif  // USE_ESP32 || (USE_ZEPHYR && USE_NRF52)
   void dump_config() override;
 
   void update() override;
