+++ esphome/components/internal_temperature/sensor.py
@@ -1,6 +1,7 @@
 import esphome.codegen as cg
 from esphome.components import sensor
 from esphome.components.zephyr import zephyr_add_prj_conf
+from esphome.config_helpers import filter_source_files_from_platform
 import esphome.config_validation as cv
 from esphome.const import (
     DEVICE_CLASS_TEMPERATURE,
@@ -11,6 +12,7 @@ from esphome.const import (
     PLATFORM_RP2040,
     STATE_CLASS_MEASUREMENT,
     UNIT_CELSIUS,
+    PlatformFramework,
 )
 from esphome.core import CORE
 
@@ -39,3 +41,18 @@ async def to_code(config):
     if CORE.using_zephyr and CORE.is_nrf52:
         zephyr_add_prj_conf("SENSOR", True)
         zephyr_add_prj_conf("TEMP_NRF5", True)
+
+
+FILTER_SOURCE_FILES = filter_source_files_from_platform(
+    {
+        "internal_temperature_esp32.cpp": {
+            PlatformFramework.ESP32_ARDUINO,
+            PlatformFramework.ESP32_IDF,
+        },
+        "internal_temperature_rp2040.cpp": {PlatformFramework.RP2040_ARDUINO},
+        "internal_temperature_bk72xx.cpp": {
+            PlatformFramework.BK72XX_ARDUINO,
+        },
+        "internal_temperature_zephyr.cpp": {PlatformFramework.NRF52_ZEPHYR},
+    }
+)
